docs: Rename additional page files so they appear in alphabetical order
doxygen scans through files in alphabetical order when a directory is
given for `INPUT`. The displayed title in "Related Pages" is not the
same as the file name, so the order does not appear correctly in the
contents. Thus the new file names contain now a positioning number to
control the order they appear in the contents.

Closes https://github.com/Makman2/CE3D2/issues/132