make.sh: Add publish target
This target performs necessary release tasks, like creating necessary
branches and commits and advancing versions. It also packs the binaries
into archives as well as the documentation.

The old `pack`-target did a part of the publishing job by building the
archives. However, the old archive structure does not match the new
desired archive content, so this target was removed altogether.

Closes https://github.com/Makman2/CE3D2/issues/136